+++ contracts/mocks/sale/FixedPricesSaleMock.sol
@@ -5,7 +5,6 @@ pragma solidity 0.6.8;
 import "../../sale/FixedPricesSale.sol";
 
 contract FixedPricesSaleMock is FixedPricesSale {
-
     /**
      * Constructor.
      * @dev Emits the `MagicValues` event.
@@ -20,4 +19,12 @@ contract FixedPricesSaleMock is FixedPricesSale {
         uint256 tokensPerSkuCapacity
     ) public FixedPricesSale(payoutWallet_, skusCapacity, tokensPerSkuCapacity) {}
 
+    function createSku(
+        bytes32 sku,
+        uint256 totalSupply,
+        uint256 maxQuantityPerPurchase,
+        address notificationsReceiver
+    ) external {
+        _createSku(sku, totalSupply, maxQuantityPerPurchase, notificationsReceiver);
+    }
 }