+++ contracts/sale/interfaces/IOracleSale.sol
@@ -2,32 +2,26 @@
 
 pragma solidity 0.6.8;
 
-
 /**
  * @title IOracleSale
- * An ISale which can manage pricing for some tokens through an oracle.
- * TODO
+ * An ISale which can manage token conversion rates through an oracle.
  */
-interface IOracleSale /*is ISale*/{
+/*is ISale*/
+interface IOracleSale {
     /**
-     * Returns the price magic value used to represent an oracle-based pricing.
+     * Returns the price magic value used to represent an oracle-based token conversion pricing.
      * @dev MUST NOT be zero. SHOULD BE a prohibitively big value, so that it doesn’t collide with a possible price value.
-     * @return the price magic value used to represent an oracle-based pricing.
+     * @return The price magic value used to represent an oracle-based token conversion pricing.
      */
+    // solhint-disable-next-line func-name-mixedcase
     function PRICE_CONVERT_VIA_ORACLE() external pure returns (uint256);
 
     /**
-     * Returns the token used as reference for oracle-based price conversions.
-     * @dev MUST NOT be the zero address.
-     * @return the token used as reference for oracle-based price conversions.
-     */
-    function referenceToken() external view returns (address);
-
-    /**
-     * Retrieves the current rates for the `referenceToken`/`tokens` pairs via the oracle.
-     * @dev Reverts if the oracle does not provide a pricing for one of the pairs.
+     * Retrieves the token conversion rates for the `tokens`/`referenceToken` pairs via the oracle.
+     * @dev Reverts if the oracle does not provide a conversion rate for one of the token pairs.
      * @param tokens The list of tokens to retrieve the conversion rates for.
-     * @return rates the rates for the `referenceToken`/`tokens` pairs retrieved via the oracle.
+     * @param data Additional data with no specified format for deriving the conversion rates.
+     * @return rates The conversion rates for the `tokens`/`referenceToken` pairs, retrieved via the oracle.
      */
-    function conversionRates(address[] calldata tokens) external view returns (uint256[] memory rates);
+    function conversionRates(address[] calldata tokens, bytes calldata data) external view returns (uint256[] memory rates);
 }