+++ test/scenarios/index.js
@@ -1,0 +1,5 @@
+const purchasing = require('./purchasing.scenario');
+
+module.exports = {
+  ...purchasing,
+};